Move application to own folder
