+++ wsgi.py
@@ -1,4 +1,4 @@
-from app import app
+from climatedata_api.app import app
 
 if __name__ == "__main__":
     app.run()
